// src/features/chart/schema/enums.gen.ts
export type CelestialBodyId = string;

export interface CelestialBodyInfo {
  displayName: string;
  glyphAsset: string;
}

export const CELESTIAL_BODIES: Record<CelestialBodyId, CelestialBodyInfo> = {
  "sun": { displayName: "Sun", glyphAsset: "sun" },
  "moon": { displayName: "Moon", glyphAsset: "moon" },
  "mercury": { displayName: "Mercury", glyphAsset: "mercury" },
  "venus": { displayName: "Venus", glyphAsset: "venus" },
  "mars": { displayName: "Mars", glyphAsset: "mars" },
  "jupiter": { displayName: "Jupiter", glyphAsset: "jupiter" },
  "saturn": { displayName: "Saturn", glyphAsset: "saturn" },
  "uranus": { displayName: "Uranus", glyphAsset: "uranus" },
  "neptune": { displayName: "Neptune", glyphAsset: "neptune" },
  "pluto": { displayName: "Pluto", glyphAsset: "pluto" },
  "ascendant": { displayName: "Ascendant", glyphAsset: "ascendant" },
  "midheaven": { displayName: "Midheaven", glyphAsset: "midheaven" },
  "descendant": { displayName: "Descendant", glyphAsset: "descendant" },
  "imumCoeli": { displayName: "Imum Coeli", glyphAsset: "imum coeli" },
  "rahu": { displayName: "North Node", glyphAsset: "north node" },
  "ketu": { displayName: "South Node", glyphAsset: "south node" },
  "fortune": { displayName: "Part of Fortune", glyphAsset: "part of fortune" },
  "spirit": { displayName: "Lot of Spirit", glyphAsset: "lot of soul" },
  "lotOfEros": { displayName: "Lot of Eros", glyphAsset: "eros" },
  "vertex": { displayName: "Vertex", glyphAsset: "vertex" },
  "blackMoonLilith": { displayName: "Black Moon Lilith", glyphAsset: "lilith" },
  "chiron": { displayName: "Chiron", glyphAsset: "chiron" },
  "ceres": { displayName: "Ceres", glyphAsset: "ceres" },
  "pallas": { displayName: "Pallas", glyphAsset: "pallas" },
  "juno": { displayName: "Juno", glyphAsset: "juno" },
  "vesta": { displayName: "Vesta", glyphAsset: "vesta" },
  "eros": { displayName: "Eros", glyphAsset: "eros-asteroid" },
  "pholus": { displayName: "Pholus", glyphAsset: "pholus" },
  "astraea": { displayName: "Astraea", glyphAsset: "astraea" },
  "hebe": { displayName: "Hebe", glyphAsset: "hebe" },
  "iris": { displayName: "Iris", glyphAsset: "iris" },
  "flora": { displayName: "Flora", glyphAsset: "flora" },
  "metis": { displayName: "Metis", glyphAsset: "metis" },
  "hygiea": { displayName: "Hygiea", glyphAsset: "hygiea" },
  "urania": { displayName: "Urania", glyphAsset: "urania" },
  "hilda": { displayName: "Hilda", glyphAsset: "hilda" },
  "philosophia": { displayName: "Philosophia", glyphAsset: "philosophia" },
  "sophia": { displayName: "Sophia", glyphAsset: "sophia" },
  "aletheia": { displayName: "Aletheia", glyphAsset: "aletheia" },
  "sapientia": { displayName: "Sapientia", glyphAsset: "sapientia" },
  "thule": { displayName: "Thule", glyphAsset: "thule" },
  "ursula": { displayName: "Ursula", glyphAsset: "ursula" },
  "hidalgo": { displayName: "Hidalgo", glyphAsset: "hidalgo" },
  "lilithAsteroid": { displayName: "Lilith (Asteroid)", glyphAsset: "lilith (asteroid)" },
  "amor": { displayName: "Amor", glyphAsset: "amor" },
  "kama": { displayName: "Kama", glyphAsset: "kama" },
  "aphrodite": { displayName: "Aphrodite", glyphAsset: "aphrodite" },
  "apollo": { displayName: "Apollo", glyphAsset: "apollo" },
  "damocles": { displayName: "Damocles", glyphAsset: "damocles" },
  "cruithne": { displayName: "Cruithne", glyphAsset: "cruithne" },
  "vulcano": { displayName: "Vulcano", glyphAsset: "vulcano" },
  "nessus": { displayName: "Nessus", glyphAsset: "nessus" },
  "cupidoUranian": { displayName: "Cupido (Uranian)", glyphAsset: "cupido (uranian)" },
  "hades": { displayName: "Hades", glyphAsset: "hades" },
  "zeusUranian": { displayName: "Zeus (Uranian)", glyphAsset: "zeus (uranian)" },
  "kronos": { displayName: "Kronos", glyphAsset: "kronos" },
  "apollon": { displayName: "Apollon", glyphAsset: "apollon" },
  "admetos": { displayName: "Admetos", glyphAsset: "admetos" },
  "vulkanus": { displayName: "Vulkanus", glyphAsset: "vulkanus" },
  "poseidonUranian": { displayName: "Poseidon (Uranian)", glyphAsset: "poseidon (uranian)" },
  "regulus": { displayName: "Regulus", glyphAsset: "regulus" },
  "aldebaran": { displayName: "Aldebaran", glyphAsset: "aldebaran" },
  "antares": { displayName: "Antares", glyphAsset: "antares" },
  "fomalhaut": { displayName: "Fomalhaut", glyphAsset: "fomalhaut" },
  "vega": { displayName: "Vega", glyphAsset: "vega" },
  "sirius": { displayName: "Sirius", glyphAsset: "sirius" },
  "altair": { displayName: "Altair", glyphAsset: "altair" },
  "capella": { displayName: "Capella", glyphAsset: "capella" },
  "rigel": { displayName: "Rigel", glyphAsset: "rigel" },
  "procyon": { displayName: "Procyon", glyphAsset: "procyon" },
  "betelgeuse": { displayName: "Betelgeuse", glyphAsset: "betelgeuse" },
  "spica": { displayName: "Spica", glyphAsset: "spica" },
  "algol": { displayName: "Algol", glyphAsset: "algol" },
};